modify : PageController
페이지 컨트롤러 통합
