+++ src/main/java/com/bringup/common/security/config/SpringSecurityConfig.java
@@ -142,6 +142,11 @@ public class SpringSecurityConfig {
                 antMatcher("/member/companyReview"),
                 antMatcher("/member/interviewReview"),
                 antMatcher("/member/notice"),
+                antMatcher("/member/recruitmentPage"),
+                antMatcher("/member/topRecruitment"),
+                antMatcher("/member/confirmRecruitment"),
+                antMatcher("/member/AnnouncementRecruitment"),
+                antMatcher("/member/proposeRecruitment"),
                 antMatcher(GET, "image/**"),
                 antMatcher("/recruitment/**"),
                 antMatcher("/main/image/**"),
