+++ src/main/webapp/WEB-INF/views/member/header/member_header.jsp
@@ -165,11 +165,11 @@
                     <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle"  id="jobPostingMenu" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">채용 정보</a>
                         <ul class="dropdown-menu" aria-labelledby="jobPostingMenu">
-                            <li><a class="dropdown-item" href="/recruitment/recruitmentPage">홈</a></li>
-                            <li><a class="dropdown-item" href="/recruitment/topRecruitment">인기 공고 TOP 100</a></li>
-                            <li><a class="dropdown-item" href="/recruitment/confirmRecruitment">내가 본 공고</a></li>
-                            <li><a class="dropdown-item" href="/recruitment/AnnouncementRecruitment">내가 지원한 공고</a></li>
-                            <li><a class="dropdown-item" href="/recruitment/proposeRecruitment">제안 받은 공고</a></li>
+                            <li><a class="dropdown-item" href="/member/recruitmentPage">홈</a></li>
+                            <li><a class="dropdown-item" href="/member/topRecruitment">인기 공고 TOP 100</a></li>
+                            <li><a class="dropdown-item" href="/member/confirmRecruitment">내가 본 공고</a></li>
+                            <li><a class="dropdown-item" href="/member/AnnouncementRecruitment">내가 지원한 공고</a></li>
+                            <li><a class="dropdown-item" href="/member/proposeRecruitment">제안 받은 공고</a></li>
                         </ul>
                     </li>
